remove duplicate files, update imports
