+++ src/apps/favorites-list/FavoritesList.tsx
@@ -4,7 +4,7 @@ import usePager from "../../components/result-pager/use-pager";
 import { useGetList } from "../../core/material-list-api/material-list";
 import { useText } from "../../core/utils/text";
 import { Pid } from "../../core/utils/types/ids";
-import CardListItemAdapter from "../../components/card-list/card-list-item-adapter";
+import CardListItemAdapter from "../../components/card-item-list/card-list-item/card-list-item-adapter";
 
 export interface FavoritesListProps {
   pageSize: number;
