+++ src/components/card-item-list/card-list-item/card-list-item-cover.tsx
@@ -1,9 +1,10 @@
 import React from "react";
 import { Cover, CoverProps } from "../../cover/cover";
 
-type SearchResultListItemCoverProps = Omit<CoverProps, "animate" | "size">;
-const SearchResultListItemCover: React.FC<SearchResultListItemCoverProps> = ({
+type CardListItemCoverProps = Omit<CoverProps, "animate" | "size">;
+const CardListItemCover: React.FC<CardListItemCoverProps> = ({
   id,
+  description,
   url,
   tint,
   linkAriaLabelledBy
@@ -13,6 +14,7 @@ const SearchResultListItemCover: React.FC<SearchResultListItemCoverProps> = ({
       animate
       id={id}
       size="small"
+      description={String(description)}
       url={url}
       tint={tint}
       linkAriaLabelledBy={linkAriaLabelledBy}
@@ -20,4 +22,4 @@ const SearchResultListItemCover: React.FC<SearchResultListItemCoverProps> = ({
   );
 };
 
-export default SearchResultListItemCover;
+export default CardListItemCover;
